Added the ability to view, update the number of items and remove items from a user's cart.
Added custom number input box when viewing the cart.

Added an item, that already exists in the user's cart, will be added to the amount that already exists in the user's cart.

Created a global  function that creates a label and input element based on the type of data passed in.